+++ api/src/main/kotlin/api/routing/lines/LinesClient.kt
@@ -1,46 +1,41 @@
 package api.routing.lines
 
-import api.db.getItinerariesByItineraryCode
+import api.db.getItinerariesByFullLineCode
+import api.db.getRoute
 import api.db.getShapesByItineraryCode
+import api.exceptions.BusTrackerException.BadRequest
 import api.exceptions.BusTrackerException.NotFound
 import api.extensions.getWrapped
-import api.routing.Response.*
-import api.routing.lines.bus.getKmlText
+import api.routing.Response.ResponseFlowJson
+import api.routing.Response.ResponseJson
+import api.routing.stops.bus.busCodMode
 import api.utils.Pipeline
 import arrow.core.continuations.either
+import crtm.utils.createStopCode
+import crtm.utils.getCodModeFromLineCode
+import crtm.utils.getSimpleLineCodeFromLineCode
 import io.ktor.http.*
 import io.ktor.http.content.*
 import io.ktor.server.application.*
 import io.ktor.server.plugins.cachingheaders.*
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.map
-import simpleJson.asJson
 
 
-suspend fun Pipeline.getItinerariesByItineraryCode() = either {
-    val itineraryCode = call.parameters.getWrapped("itineraryCode").bind()
+suspend fun Pipeline.getItineraries(codMode: String) = either {
+    val lineCode = call.parameters.getWrapped("lineCode").bind()
+    val direction = call.parameters.getWrapped("direction").bind().toIntOrNull() ?: shift<Nothing>(BadRequest())
+    val stopCode = call.request.queryParameters.getWrapped("stopCode").bind()
 
-    val itineraries =
-        getItinerariesByItineraryCode(itineraryCode).firstOrNull() ?: shift<Nothing>(NotFound("Itinerary not found"))
+    val itinerary = getItinerariesByFullLineCode(lineCode, direction, createStopCode(codMode, stopCode)).firstOrNull()
+        ?: shift<Nothing>(NotFound())
 
-    val itinerariesOrdered =
-        itineraries.copy(stops = itineraries.stops.distinctBy { it.fullStopCode to it.order }.sortedBy { it.order })
+    val itineraryOrdered =
+        itinerary.copy(stops = itinerary.stops.distinctBy { it.fullStopCode to it.order }.sortedBy { it.order })
 
-    val json = itinerariesOrdered.let(::buildItineraryJson).asJson()
+    val json = itineraryOrdered.let(::buildItineraryJson)
     call.caching = CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 60 * 60))
     ResponseJson(json, HttpStatusCode.OK)
-}
-
-suspend fun Pipeline.getKml() = either {
-    val itineraryCode = call.parameters.getWrapped("itineraryCode").bind()
-
-    val itinerary =
-        getItinerariesByItineraryCode(itineraryCode).firstOrNull() ?: shift<Nothing>(NotFound("Itinerary not found"))
-
-    val kml = getKmlText(itinerary).bind()
-
-    call.caching = CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 24 * 60))
-    ResponseString(kml, contentType = ContentType.parse("text/xml"), HttpStatusCode.OK)
 }
 
 suspend fun Pipeline.getShapes() = either {
@@ -51,4 +46,32 @@ suspend fun Pipeline.getShapes() = either {
     val json = shapes.map(::buildShapeJson)
     call.caching = CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 60 * 60))
     ResponseFlowJson(json, HttpStatusCode.OK)
+}
+
+suspend fun Pipeline.getLocations() = either {
+    val lineCode = call.parameters.getWrapped("lineCode").bind()
+    val direction = call.parameters.getWrapped("direction").bind().toIntOrNull() ?: shift<Nothing>(BadRequest())
+    val stopCode = call.request.queryParameters.getWrapped("stopCode").bind()
+
+    val fullStopCode = createStopCode(busCodMode, stopCode)
+    val codMode = getCodModeFromLineCode(lineCode)
+    val route = getRoute(lineCode).getOrNull()
+    val simpleLineCode = route?.simpleLineCode ?: getSimpleLineCodeFromLineCode(lineCode)
+    val routeCodMode = route?.codMode ?: busCodMode
+
+    val locations = VehicleLocations(
+        locations = getLocationsResponse(
+            lineCode,
+            direction,
+            codMode,
+            fullStopCode
+        ).bind().vehiclesLocation.vehicleLocation,
+        codMode = routeCodMode.toInt(),
+        lineCode = simpleLineCode,
+    )
+
+    val json = buildVehicleLocationJson(locations)
+
+    call.caching = CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 10))
+    ResponseJson(json, HttpStatusCode.OK)
 }