+++ api/src/main/kotlin/api/routing/lines/bus/BusLinesRouting.kt
@@ -1,27 +1,10 @@
 package api.routing.lines.bus
 
-import api.extensions.handle
-import api.routing.lines.getItinerariesByItineraryCode
-import api.routing.lines.getKml
-import api.routing.lines.getShapes
+import api.routing.lines.linesConfigF
+import api.routing.stops.bus.busCodMode
 import io.ktor.server.routing.*
 
 fun Route.busLinesRouting() = route("/bus") {
-
-    get("/shapes/{itineraryCode}") {
-        handle { getShapes() }
-    }
-
-    get("/locations/{itineraryCode}") {
-        handle { getLocationsByItineraryCode() }
-    }
-
-    get("/itineraries/{itineraryCode}") {
-        handle { getItinerariesByItineraryCode() }
-    }
-
-    get("/kml/{itineraryCode}") {
-        handle { getKml() }
-    }
+    linesConfigF(busCodMode)
 }