+++ api/src/main/kotlin/api/routing/lines/emt/EmtLinesRouting.kt
@@ -1,26 +1,9 @@
 package api.routing.lines.emt
 
-import api.extensions.handle
-import api.routing.lines.getItinerariesByItineraryCode
-import api.routing.lines.getKml
-import api.routing.lines.getShapes
+import api.routing.lines.linesConfigF
+import api.routing.stops.emt.emtCodMode
 import io.ktor.server.routing.*
 
 fun Route.emtLinesRouting() = route("/emt") {
-
-    get("/shapes/{itineraryCode}") {
-        handle { getShapes() }
-    }
-
-    get("/locations/{itineraryCode}") {
-        handle { getLocationsByItineraryCode() }
-    }
-
-    get("/itineraries/{itineraryCode}") {
-        handle { getItinerariesByItineraryCode() }
-    }
-
-    get("/kml/{itineraryCode}") {
-        handle { getKml() }
-    }
+    linesConfigF(emtCodMode)
 }